refactor: 예약 스케줄과 예약 일시 분리
