+++ src/test/java/roomescape/domain/reservation/ReservationDateTimeTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static roomescape.DateUtils.yesterday;
@@ -7,10 +7,11 @@ import static roomescape.TestFixtures.anyThemeWithNewId;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.RoomescapeSchedule;
 import roomescape.domain.timeslot.TimeSlot;
 import roomescape.exception.BusinessRuleViolationException;
 
-class RoomescapeScheduleTest {
+class ReservationDateTimeTest {
 
     @Test
     @DisplayName("예약하려는 방탈출 일정의 일시가 현재 일시보다 이전이면 예외가 발생한다.")
